Create ping_system.py
This is a script request made by Overforyou#9290 and I thought this is good enough to be here.
This library lets you check a player's ping with an event-based system. The same system is used to know if a user is a bot.
If you apply tfm.exec.lowerSyncDelay over a player it will be more accurate and if you use timers it will even better (remember that you can't synchronize player's data in less than 400ms (with lowerSyncDelay) or 500ms (without it) so it wouldn't make sense to run ping.checkPings() every 25ms, but it would running it, lets say, every 333ms)
You'd need to run ping.loadGrounds() every new map so the system doesn't fail, and the given position must be an empty place of (at least) 25x25 px, with the given position (of ping.setCheckPosition) right in the middle of it.